rename subbgenerators templates directories
