+++ src/assets/templates/generate-swap-generator-main/example.txt
