+++ src/assets/templates/generate-swap-generator-main/generator-root.js
