+++ src/assets/templates/generate-swap-generator-main/generator-src.js
