+++ src/assets/templates/generate-swap-generator-main/generator.test.js
