+++ src/assets/templates/generate-swap-generator-main/index-root.js
