+++ src/assets/templates/generate-swap-generator-main/plugin.test.js
