+++ src/assets/templates/generate-swap-generator-package/$package.json
