+++ src/assets/templates/generate-swap-generator-package/gulp-tasks/release.js
