+++ src/lib/subgenerators/generate-swap-generator-package/generator.js
